# final-project/docs/swagger.yaml
definitions:
...
  dto.SocialMediaCreate:
    properties:
      name:
        example: Twitter
        type: string
      social_media_url:
        example: https://twitter.com/budiganteng
        type: string
    type: object
  dto.SocialMediaCreateResponse:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      social_media_url:
        type: string
      user_id:
        type: integer
    type: object
  dto.SocialMediaResponse:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      social_media_url:
        type: string
      updated_at:
        type: string
      user:
        $ref: '#/definitions/dto.User'
      user_id:
        type: integer
    type: object
  dto.SocialMediaUpdate:
    properties:
      name:
        example: X
        type: string
      social_media_url:
        example: https://x.com/budiganteng
        type: string
    type: object
  dto.SocialMediaUpdateResponse:
    properties:
      id:
        type: integer
      name:
        type: string
      social_media_url:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  dto.User:
    properties:
      email:
        type: string
      username:
        type: string
    type: object
  dto.UserCreateResponse:
    properties:
      age:
        type: integer
      email:
        type: string
      id:
        type: integer
      username:
        type: string
    type: object
  dto.UserLogin:
    properties:
      email:
        example: budiganteng@rocketmail.com
        type: string
      password:
        example: budiganteng123
        type: string
    type: object
  dto.UserLoginResponse:
    properties:
      token:
        type: string
    type: object
  dto.UserRegister:
    properties:
      age:
        example: 20
        type: integer
      email:
        example: budiganteng@rocketmail.com
        type: string
      password:
        example: budiganteng123
        type: string
      username:
        example: budiganteng
        type: string
    type: object
  dto.UserUpdate:
    properties:
      email:
        example: budigantengbanget@rocketmail.com
        type: string
      username:
        example: budigantengbanget
        type: string
    type: object
  dto.UserUpdateResponse:
    properties:
      age:
        type: integer
      email:
        type: string
      id:
        type: integer
      updated_at:
        type: string
      username:
        type: string
    type: object
  helper.Response-any:
    properties:
      data: {}
      errors:
        items:
          type: string
        type: array
      message:
        type: string
      success:
        type: boolean
    type: object
...
  helper.Response-array_dto_SocialMediaResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.SocialMediaResponse'
        type: array
      errors:
        items:
          type: string
        type: array
      message:
        type: string
      success:
        type: boolean
    type: object
...
  helper.Response-dto_SocialMediaCreateResponse:
    properties:
      data:
        $ref: '#/definitions/dto.SocialMediaCreateResponse'
      errors:
        items:
          type: string
        type: array
      message:
        type: string
      success:
        type: boolean
    type: object
  helper.Response-dto_SocialMediaResponse:
    properties:
      data:
        $ref: '#/definitions/dto.SocialMediaResponse'
      errors:
        items:
          type: string
        type: array
      message:
        type: string
      success:
        type: boolean
    type: object
  helper.Response-dto_SocialMediaUpdateResponse:
    properties:
      data:
        $ref: '#/definitions/dto.SocialMediaUpdateResponse'
      errors:
        items:
          type: string
        type: array
      message:
        type: string
      success:
        type: boolean
    type: object
  helper.Response-dto_UserCreateResponse:
    properties:
      data:
        $ref: '#/definitions/dto.UserCreateResponse'
      errors:
        items:
          type: string
        type: array
      message:
        type: string
      success:
        type: boolean
    type: object
  helper.Response-dto_UserLoginResponse:
    properties:
      data:
        $ref: '#/definitions/dto.UserLoginResponse'
      errors:
        items:
          type: string
        type: array
      message:
        type: string
      success:
        type: boolean
    type: object
  helper.Response-dto_UserUpdateResponse:
    properties:
      data:
        $ref: '#/definitions/dto.UserUpdateResponse'
      errors:
        items:
          type: string
        type: array
      message:
        type: string
      success:
        type: boolean
    type: object
info:
  contact: {}
  description: submission for final-project
  title: Hacktiv8-Golang final-project
  version: "1.0"
paths:
...
  /socialmedias:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.Response-array_dto_SocialMediaResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.Response-any'
      security:
      - BearerToken: []
      summary: get all social media
      tags:
      - social-media
    post:
      consumes:
      - application/json
      parameters:
      - description: required body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.SocialMediaCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/helper.Response-dto_SocialMediaCreateResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.Response-any'
      security:
      - BearerToken: []
      summary: create a new social media
      tags:
      - social-media
  /socialmedias/{socialMediaID}:
    delete:
      parameters:
      - description: social media ID
        in: path
        name: socialMediaID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.Response-any'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Response-any'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.Response-any'
      security:
      - BearerToken: []
      summary: delete social media
      tags:
      - social-media
    get:
      parameters:
      - description: social media ID
        in: path
        name: socialMediaID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.Response-dto_SocialMediaResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Response-any'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.Response-any'
      security:
      - BearerToken: []
      summary: get social media by ID
      tags:
      - social-media
    put:
      consumes:
      - application/json
      parameters:
      - description: social media ID
        in: path
        name: socialMediaID
        required: true
        type: integer
      - description: required body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.SocialMediaUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.Response-dto_SocialMediaUpdateResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Response-any'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.Response-any'
      security:
      - BearerToken: []
      summary: update social media
      tags:
      - social-media
  /users:
    delete:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.Response-any'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.Response-any'
      security:
      - BearerToken: []
      summary: delete user
      tags:
      - users
    put:
      consumes:
      - application/json
      parameters:
      - description: required body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UserUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.Response-dto_UserUpdateResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.Response-any'
      security:
      - BearerToken: []
      summary: update user
      tags:
      - users
  /users/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: required body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UserLogin'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.Response-dto_UserLoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.Response-any'
      summary: login user
      tags:
      - users
  /users/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: required body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UserRegister'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/helper.Response-dto_UserCreateResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.Response-any'
      summary: register a new user
      tags:
      - users
securityDefinitions:
  BearerToken:
    description: 'Bearer token for authentication. Example: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"